Bug#13354910 - MAKE_JOIN_STATISTICS() INVOKES                RECORDS_IN_RANGE(MIN_KEY=0X0, MAX_KEY=0X0)
Some range predicates that by them selves make up partial 
ranges are merged into a full range by the range optimizer. 
An example is the query for this bug:

  "WHERE col IN (1,2,6) OR col NOT IN (1)"

This predicate is always true and is correctly merged into
the range [-inf <= col <= +inf]. However, although the range
optimizer did make a full range, it did not realize that it's
meaningless to use this as a range predicate since that will 
be the same as doing an index scan.

For the query in the bug the problem was that 
SEL_ARG::copy_max() had a bug: When checking if the range was
a full range, it was checked if max_flag was set to 
(NO_MAX_RANGE | NO_MIN_RANGE). But the NO_MIN_FLAG is stored
in min_flag, so the test is changed to 
(max_flag & NO_MAX_FLAG) and (min_flag & NO_MIN_FLAG).

Enabling the DBUG_ASSERT(max_key || min_key) in 
ha_innobase::records_in_range() then showed another related 
bug: innodb_icp started failing because SEL_ARG::copy_min()
had the very same bug as copy_max(). This was fixed as well.

In an attempt at unittesting SEL_ARG, copy_min() and copy_max()
have now been added to the opt_range-t gunit test.

To see the assert, add "DBUG_ASSERT(min_key || max_key);"
to the beginning of ha_innobase::records_in_range().