Bug #13628249 - ASSERT MAX_UNDO_LOGS == ULINT_UNDEFINED WHEN INNODB-FORCE-RECOVERY >= 5 Bug #13628420	NORML SHUTDOWN HANGS WHEN INNODB-FORCE-RECOVERY=3 AND THERE ARE TXNS TO ROLLBACK
There are two parts to this fix, the first was introduced by the fix for
bug#13255225. The srv_undo_logs should be set to ULINT_UNDEFINED if force
recovery >= 3.

The second fix is to prevent data dictionary temp tables and indexes drop,
if force recovery >= 3. This is the same level at which we prevent rollback
of recovered user transactions.

rb://944 Approved by Jimmy Yang.