Fix for Bug#13655397 CRASH IN SYNC_THREAD_LEVELS_NONEMPTY_TRX
This crash occured due to pushing an index condition down to InnoDB
that contained a call to a stored function. When the index condition
was executed from within InnoDB, the stored function caused a second
call from the server into InnoDB. Since InnoDB does not handle that
the same session executes more than one operation against InnoDB
at a given time this triggered an assert in InnoDB.

The fix is to prevent that we push down index conditions to the
storage engine that contain calls to stored functions.