BUG#13728026 - BINLOG MAY CAUSES CRASHES WHEN THE SLAVE'S REPOSITORY IS A TABLE.
The command BINLOG requires a Relay_log_info object to execute events.
      
Such object was being created with INFO_REPOSITORY_FILE causing issues
when the main Relay_log_info object used INFO_REPOSITORY_TABLE. Among
a variety of issues, this led to the existence of two different types
of repositories thus making new threads to fail when they tried to
create a new Relay_log_info.
      
The following message was usually printed out: 
      
"Error creating relay log info: Multiple replication metadata repository
instances found with data in them. Unable to decide which is the correct
one to choose."
      
To fix the problem we create this particular Relay_log_info object with
a INFO_REPOSITORY_DUMMY as in fact a repository is not necessary. We
also improved the code for the case rli is NULL. 
In the future, we need to improve the code around the BINLOG command as
only a small part of the object is required to execute it.