Bug#13727933 - PURGE BINARY LOGS TO MIGHT BE INEFFECTIVE AND NOT VERBOSE ABOUT THAT
This is a workaround patch as PURGE can't be used so RESET MASTER is attempted as a tool 
to remove binlog files and thereby to desynchronize master and slave wrt GTID data.

I had to change the test in three more parts: two are the local slave side
INSERT and DROP as parts of the workaround, and
yet another 
--source include/have_binlog_format_mixed_or_statement.inc
due to a new being reported Bug 13729084.