Bug #13628249 - ASSERT MAX_UNDO_LOGS == ULINT_UNDEFINED WHEN INNODB-FORCE-RECOVERY >= 5
Introduce ULONG_UNDEFINED, on 64 bit Windows sizeof(ulong) < sizeof(ulint). When
--innodb-force-recovery >= 3 we set srv_undo_logs to ULINT_UNDEFINED and then
later when we go to start a transaction and attach a rollback segment in the
function trx_assign_rseg(), only the lower 32 bits are copied because the
type of the function parameter is a ulint. This causes the assertion to
fail because it is comparing it with ULINT_UNDEFINED which is 64 bits. The fix
is to change the type of the parameter to ulong and introduce ULONG_UNDEFINED
and additionally check against that instead of ULINT_UNDEFINED in the assert.

Approved by Jimmy Yang over IM.