BUG#13729084 - COMBINATION OF RESET MASTER AND ROW FORMAT LEADS TO EVENT APPLY RATHER THAN SKIP
The function gtid_pre_statement_checks() checks if a transaction
was already processed and thus has to be skipped or otherwise has
to be processed.

In row mode, this routine was returning that already processed DML
changes should be processed. This happened because the sql_command
was not being reset and when a row event was procossed the function
gtid_pre_statement_checks() got sql_command set to SQLCOM_BEGIN.

To fix the problem, we reset sql_command to SQLCOM_END, as soon as
a Query event is processed.