[part 2 of WL5655] Add vectored IO support to innodb
rb://895
approved by: Marko Makela

Currently we memcpy data pages to a buffer and issue a single IO from
there. This patch eliminates that memcpy stuff and instead use vio
support where present. Note that I have not tried to tidy up the dblwr
interface with vio as we are going to revisit this in next part of the
patch.

The Implementation:
===================
* On windows use scatter/gather routines. For now disabled.
* On non-windows platforms use one of the following in that order:
  * preadv/pwritev where available. Need newer kernel version.
  * readv/writev: should be available on all platforms.
    Need to do an lseek() and then lock the pointer before issuing
    a readv/writev.
  * pread/pread: In rare cases where readv/writev is not available we
    fall back to copying to a buffer and using pread/pwrite
  * read/write: In some ancient platform which doesn't even support
    pread/pwrite we fall back to read/write.

Limitations:
============
Some arbitrary limitations in the current implementation:
* A given vio struct will support 1000 * UNIV_PAGE_SIZE batch maximum.
* IO requests are to be multiple of UNIV_PAGE_SIZE
* Windows scatter/gather routines are complicated in that they don't
  take length as parameter. The code for scatter/gather is in place but
  is disabled for now.

Callers:
========
Current caller of vio routine is just dblwr. The support for read is
also present but is not currently used. For all batch flushing vio is
used to write to the doublewrite buffer. For single page flushes when
writing uncompressed page it is written to the dblwr directly using
fil_io. When writing compressed page vio is used with two IO requests,
first request is the page itself and the second request is the zero
padding.