WL#3584
The rpl_gtid_stress_failover fails in MTS mode with messages:

"1201: Could not initialize master info structure; more error 
messages can be found in the MySQL error log"

and 

"Error looking for file after ./mysqld-relay-bin.000004."

These are the symptoms of BUG#12995174 in mysql-trunk. Disabling 
the test case in MTS mode until this is fixed in mysql-trunk.