Bug#13571700 TINYBLOB NOT NULL, CRASH IN PROTOCOL::NET_STORE_DATA
Problem: Crash or wrong result for NOT NULL columns if the query
contains an impossible WHERE clause.

If a query contains an impossible WHERE clause, JOIN::optimize()
recognizes that this query will not produce any rows (except maybe a
NULL row) and exits early, without constructing join_tab.

JOIN::exec() recognizes the optimizer's decision and calls
return_zero_rows() to construct the NULL row. To construct this row,
it first calls mark_as_null_row() on the tables and then reads the
fields. However, the tables are found by following pointers from
join_tab, which is a null pointer because of the impossible WHERE
clause. The result is that the tables' null_row are not set, which
again leads to Field::is_null() returning false and an attempt to read
the field's value. The value is undefined, so the result may be an
arbitrary value (e.g., in the case of an INT) or a crash if it is an
undefined pointer, as in the case of TINYBLOB.

This bug is a regression introduced by
revid:jorgen.loland@oracle.com-20111220120431-pxmaax8e3effzhj0.

Fix: Revert to accessing tables through join->select_lex->leaf_tables
instead of join->join_tab.