BUG#13738296 - INSUFFICIENT SIZE OF MAX_ALLOWED_PACKET ON MASTER CAUSING SLAVE TO FAILED ON WIN
Transactions' changes are gathered in caches which upon commit are
written to the binary log. This cache however does not support seek
operations and an attempt to issue a seek operation against it may
corrupt the binary log.

This happens when the content of the cache does not fit in-memory
and a swap are is used. In this case, when a seek is issued the
content of the swap are is not read back to memory so that any
attempt to write the in-memory part back to the swap are will
write garbage into it.

In the WL#3584, we used a seek to reposition the cursor on the file
and rewrite the GTID thus leading to a corrupted binary log.

To fix the problem, we flush the in-memory part of the cache to the
swap area, call reinit_io_cache to reposition the cursor and read
the correct content back to memory, update the gtid and flush the
memory back to disk. This solution though has a performance drawback
as a flush to the swap area is required even when there is enough
memory to accommodate all transaction's changes.