WL#3584
Fix for rpl_gtid_mode.test wich fails in innodb_page_size=4K and in
innodb_page_size=8K collections. 

The test stops and starts the server manually, but it does not set
the innodb_page_size correctly. Thus when restarting the server without
specifying the page size that was used before stopping the server,
makes the test case fail when the page size was different from the 
default value.

We fix this by including in the options passed to the mysqld, at restart
time, the original value of innodb_page_size.