BUG#13681205: REPLICATION FAILS DURING SLAVE WORKERS INITIALIZATION,               CRASHES SLAVE
The maximum set for the option slave_checkpoint_group was off by 7.
Since this field, when stored in a table, is stored as a blob, then
its value must be 1 < slave_checkpoint_group < 2^16*8. However it was
wrongly set to 2^16*8-1 (524287) instead of 2^16*8-8 (524280).

Furthermore, when the slave failed to initialize the info repository,
it would crash sometimes because the worker jobs queue might point to
some random memory. 

We fix this by correcting the MTS_MAX_BITS_IN_GROUP and by freeing the
jobs queue only if it is flagged as initialized.