BUG#13731380 - RANGE OPTIMIZER CALLS RECORDS_IN_RANGE()                FOR OPEN RANGE
Yet another case where the range optimizer did not realize 
that the ORed range was open, i.e. always true. The case in 
bug the query contained predicates:

"WHERE col <> 3 or col < 4"

The fix is to modify key_or() to return NULL for this case
as well. See related bug 13354910.