Bug#13819132 BROKEN SYNTAX ACCEPTED FOR START TRANSACTION
Before this fix, the parser accepted the following syntax
for START TRANSACTION :

START TRANSACTION ,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,, ;

The root cause was the bison grammar for rule start_transaction_opt_list,
which produced a list of valid options or (empty string),
separated by ",".

The grammar has been fixed so that the grammar rule is:
- either en empty string
- or a list of valid options separated by ",".
which is the proper syntax.

Grammar rules have been renamed for code clarity.