Addendum to the fix for bug #13799126:
Since converting const to non-const appears to
be triggering compilation errors on some platform
made the parameter to the static function a boolean
instead of an SSL type and depending on the 
value of the boolean added a call to the appropriate
function inline.
With this the need to know the const-abiliy of this
particular function no longer exists.