Bug#54854 Can't find good position for replication break between DDL statements
Part 1 of patches to mysql-trunk.

Overall fix has two parts :
  1) Improve Ndb replication channel cutover with precise 
     next-position-after-epoch-trans in the mysql.ndb_binlog_index
     table.
  2) Simplify configuration of DDL-idempotency with a new
     MySQL Server option

Part 1) of the fix includes patches affecting the generic MySQL Server
(this commit) and patches affecting the Ndb storage engine.

Part 2) of the fix includes patches affecting the generic MySQL Server.

This commit adds a THD member which records the next Binlog position
(filename and offset) after the last Binlog event recorded by the
thread.
This information is exposed via an extension to the Binlog injector
Api.

Further patches to be pushed to the MySQL Cluster branches make use
of this functionality to fix bug#54854.

In mysql-trunk builds, this functionality has no externally visible
effect, and therefore no specific testcase.

Testcases are added in branches merging this branch and MySQL Cluster.