BUG#13820370 - RPL_GTID_MODE FAILS IN PB
The test case rpl_gtid_mode.test is doing the following steps to
check whether the server fails while starting up as supposed to be
or not:
  
  . starting slave with gtid-mode=ON; 
  . issuing "CHANGE MASTER TO AUTO_POSITION=1"; 
  . shutting down slave; 
  . trying to restart slave with gtid-mode=OFF. 
  
However, when the slave fails in the startup routine, it tries to
print out an error message with ER(X) which uses "current_thread". 
This variable is not defined though as there is no thread causing
a segmentation fault.
  
To fix the problem, we replace ER(X) by my_error().