ndb - dbacc - use DynArr256 for page directory
this will relax one size limitation of hash index
of one partition, that is the number of rows are
now not limited to ~45M, but note where are other
limitations that can be more restrictive.

pages are still from static pool (IndexMemory).
localkey length into dbtup are still only one word.