BUG#13504281 - FOLLOW UP 'LIVE UPGRADE FAILS BETWEEN 5.5 -> 5.6.4 (BUG#13427444)
Replication objects are always created so repositories are checked in the
initialization routine even when --server-id is equal to "0" (i.e. default
value), although a valid --server-id (i.e. not equal to the default value)
is a requirement to use either a master or slave.

This may cause confusion in live upgrades when an user tries to migrate
from 5.5 to 5.6 because the operation stops with an error and asking users
to provide --skip-slave-start in order to avoid the problem.

To avoid this scenario, we only create the replication objects when a valid
--server-id is provided.