BUG#13701206: WHERE A>=B DOES NOT GIVE SAME EXECUTION PLAN               AS WHERE B<=A (RANGE OPTIMIZER)
One would expect the optimizer to create the same QEP from
for these two queries:

1) SELECT * FROM t1 STRAIGHT_JOIN t2 WHERE t1.a=t2.b;
2) SELECT * FROM t1 STRAIGHT_JOIN t2 WHERE t2.b=t1.a;

However, the range optimizer did not treat (a OP b) equal to
(b OP a). Because of this, dynamic range ("Range
checked for each record") was not a possible a access method 
for query 1).

The fix is to treat "a OP b" equal to "b OP a" in the range
optimizer.