Follow-up patch for bug#13070308.
After the patch that fixes the original bug had been pushed
the xa.test began to fail sporadically with the following
diagnostics:
  query 'reap' failed with wrong errno 1205: 'Lock wait 
  timeout exceeded; try restarting transaction', instead of 1213...

It seems that the reason for such sporadically failure is a race condition
in checking for 'Sending data' state in wait_condition loop and acquiring
a real transaction lock during execution of statement 'INSERT SELECT'.
Such race condition arised since there was a time window between the step
when server sets a session state in 'Sending data' value and when the real lock
for transation is acquired. Such race condition could result in successful
execution of the following statement 'DELETE' that else would be failed with
error ER_LOCK_DEADLOCK. As a consequences the XA COMMIT for distributed
tarnsaction doesn't return any info about deadlock condition arised.

To eliminate this race condition we replaced the conditon of waiting for
lock by polling the table information_schema.innodb_trx until transaction will
be locked.