BUG#13868465 - ER_AUTO_POSITION_REQUIRES_GTID_MODE_ON PRINTS OUT A WRONG MESSAGE
"CHANGE MASTER TO" accepts the option MASTER_AUTO_POSITION in order to
enable auto_position when gtid is on. Howerver, if one tries to issue
"CHANGE MASTER TO MASTER_AUTO_POSITION" when gtid is off, the following
error message is printed out:

CHANGE MASTER TO AUTO_POSITION = 1 can only be executed when GTID_MODE = ON.

There is a typo in the error message as AUTO_POSITION does not exit.