bug#13714648 MYSQL CLUSTER ALTER TABLE REORGANIZE PARTITION UNDEFINED BEHAVIOR
Background

ALTER ONLINE TABLE REORGANIZE PARTITION (ATRP) is used to create new table partitions
(fragments), usually after new empty nodes have been added to a cluster.

ATRP determines the number of fragments to create automatically, and if
it determines that no new fragments are required, then no new fragments
will be created.

Tables are sometimes created with the MAX_ROWS option set.  This can 
be used to indicate to Ndb that extra fragments should be created to
store the large number of rows.

If MAX_ROWS was specified when a table was created, then ATRP will
use it to decide the number of partitions required, which will be the
same, so no new fragments will be created.  The table will not be 
rebalanced, and the new nodes remain empty.

One option considered was to scale the configured MAX_ROWs with the
relative increase in the number of nodes, but this results in different
numbers of partitions depending on when a table is created.

Solution

The solution implemented here is to support 
ALTER ONLINE TABLE MAX_ROWS=<bigger_value> (ATMR).  Where the bigger
MAX_ROWS value implies that more fragments are required, these will
be allocated on the new data nodes, restoring table balance.

For tables with MAX_ROWS explicitly defined, ATRP does not work.

ndb_add_partition and ndb_addnode are modified to test the new 
mechanism.