SPJ: Implement native support of pushed queries being sorted-scan-scan query.
Until now we didn't allow a scan-scan query which required 'sorted' result
set to be pushed to the SPJ block. Instead we supressed the 'sorted' request
and instead dumped the resultset to a temporary file and 'filesorted' it.
(Explained as 'Using tempfile; Using filesort')

This has been commented as 'questionable as best' by Evgeny which is
doing the review if WL5940: 'Integrating pushed join with optimizer & handler interface'
So we want to change it.....

This fix implement native support of sorted scan-scan by setting the 
parent-scan batchsize to '1' when a sorted-scan-scan request is
about to be sent to the SPJ block. 
(Dependant child scan/lookups are retreived with 'normal' batchsize)

SPJ API has to NEXTREQ more result until all related child-scan result for
the single parent row has been retrieved - This will guarantee sorted results
as all child rows will be retrieved together with its related parent.
(at the cost of some overhead though).