Fix Bug#13838962 STATISTICS CALCULATION SKIPS DELETE-MARKED RECORDS AND CAUSES ASSERTION FAILURE
Do not skip delete marked records because if all the level contains only
such records then dict_stats_analyze_index_level() will return that there
are 0 records on that level and if the level is >0 the code will assert
(only possible to have 0 records on some level if it is level 0 and the
whole tree is empty). Another possible fix is to change the code to expect
a return value of 0 for total number records and to act accordingly -
knowing that the tree is empty.

It is not clear why I added the skipping of delete-marked records in the
following changeset which this patch reverts:

* revision-id: vasil.dimov@oracle.com-20100421152317-i3m8tumk6lfr759w
* committer: Vasil Dimov <vasil.dimov@oracle.com>
* timestamp: Wed 2010-04-21 18:23:17 +0300
* message:
*   Ignore delete marked records when counting the distinct records.
*
*   If ANALYZE is run quickly after DELETE then purge may not have
*   kicked yet and delete marked records may exist.

Use TRUNCATE TABLE instead of DELETE FROM in the innodb_stats test because
with this change ANALYZE immediately after truncate becomes
non-deterministic - it is not clear how many delete-marked rows purge will
have wiped away before ANALZYE.

Approved by:	Marko (http://bur03.no.oracle.com/rb/r/993/)