Bug#13848789: SEGFAULT IN JOIN_READ_NEXT_SAME AT               SQL/SQL_EXECUTOR.CC ON HAVING...IN...JOIN
Consider a query of the form

SELECT ... HAVING (SELECT ...)

in which there is a row in the outer query that satisfies
all conditions except those in the HAVING clause. In this 
case, JOIN::exec() used to evaluate the HAVING clause and
then call return_zero_rows(). return_zero_rows() would, in
turn, call join->join_free(). Since the subquery in the HAVING
clause had already been evaluated, join_free() would assume 
that it would not be evaluated over again and would clean it 
up. However, return_zero_rows() also evaluated the HAVING 
clause, and this causes a crash. 

The fix is to not call return_zero_rows() if the reason for
no matching rows is that the HAVING clause evaluates to false.