Optimizer-trace-only changes: Fix for BUG#13799348 MAKE OPTIMIZER TRACE LESS/MORE VERBOSE REGARDING DATABASE / PLAN SEARCH
1) when greedy search is adding a table to the plan,
print the plan prefix so far (i.e. previous tables);
this is to make it easier for the reader to figure out
what the current plan is (it can get difficult
otherwise, as explained in the commit comments of
BUG 13685026)
2) for a table, instead of printing two properties in some opt trace objects
(like in "best_access_path"):
  "database": "db",
  "table": "tbl",
print one: either as
  "table": "db.tbl"
or as "db.tbl" when it's clear that it's a table (as in the "plan prefix")
this allows to remove some objects, thus decreases nesting level.
3) don't print a table's database if this database is the same as
the connection's database (which is, in practice, often the case when
we debug); this should provide more compact and readable traces.