More WL5940 addaption of ORDER/GROUP BY handling in order to address comments from reviewers:
In the case of a 'simple'(*) order by expression which can not
be executed by reading from an ordered index, there is a conflict of
interests between the optimizer and the pushed join integration:

For any 'simple' order by expression the optimizer will establish
the correct ordering of the first non-const table before starting the 
join. If there is no suitable ordered index, the table itself will be
filesorted into an intermediate record store. As the table is now
stored outside NDB, such a presorted table can not pe part of a 
pushed join.

Before this fix this conflict was resolved in favour of maximizing
usage of pushed joins: The Query Execution Plan (QEP) from the optimizer was
changes such that the presort was dropped. Instead the entire resultset
was calculated (with pushed joins) and then written to a temporary file
which was later filesorted.

The reviewer commented that changing the QEP (above) as part of the pushed join
integration was 'questionable, at best'.

This fix changes this by extending our AQP (Abstract Query Plan) with
functionality to detect when the optimizer has decided to do 
'filesort_before_join'. If that condition is detected, this presorted 
table is excluded as non-pushable.

To sum up:

- We don't any longer modify QEP as part of pushed join integration.
- This comes at the cost of some order/group by queries being less pushable.

(*) A 'simple' ordering expression contain only column references to the
    first non-const table in the join_tab's.