WL#6255 Online ALTER TABLE…MODIFY…NOT NULL, ADD FOREIGN KEY
Pre-requisite patch.

This patch splits the Alter_inplace_info handler flag
ALTER_COLUMN_NULLABLE in two - ALTER_COLUMN_NULLABLE
for changing a NOT NULL column to NULL, and
ALTER_COLUMN_NOT_NULLABLE for changing a NULL column
to NOT NULL. This is done to make it easier for storage
engines as these two operations have different requirements.