WL#5597
Fix for the case when row image is minimal and hashing is done on
an image that contains NULL fields that were not signaled in the
read_set. This means that the unpack_row function would not have 
set the null bit (it not meaningful). Ultimately, this means that
when hashing the record got from the SE, one would get different 
hash value, because the SE would have the null_bytes fully set,
whereas the BI would not.

We fix this by only checksumming the null_bytes if the read_set is
entirely set. This is the M.O. already followed by the record
comparison and the rest of the hashing calculation.