Bug#11940249: RBR: MYISAM TABLE CORRUPTION AFTER FIRST LARGE INSERT                            ON SLAVE
Problem was during repair of myisam (as a part of bulk insert optimization)
it unlocked the table, which results in when the statement used multiple
rpl row events for its new rows,  the following events was not properly
locked.

Fixed by not taking/release locks in ha_myisam::repair if already
holding a write lock.

Updated with reviewers comments.